Generalized SignUpUser Logic to SignUp
