+++ src/pages/index.js
@@ -1,9 +1,11 @@
 import FrontPage from "./FrontPage";
 import LogIn from "./LogIn";
 import LogOut from "./LogOut";
+import SignUp from './SignUp'
 
 export {
     FrontPage,
     LogIn,
-    LogOut
+    LogOut,
+    SignUp
 }
